Implement InMemoryCatalogue
The simplest way to implement this catalogue at the moment is just to
store one product. We know that it is cheating, but we don't care as
it is a fake implementation. That said, it is still important to provide
a clear message in the future if someone will attempt to use it for more
than one product.